api/pyproject.toml
[tool.poetry]
name = "GainSystemAPI"
version = "0.1.0"
description = "Packeges for GainSystemAPI"
authors = ["ronnycray <chayka1998A@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = {extras = ["debug-server"], version = "^0.81.0"}
alembic = "^1.7.1"
PyJWT = "^2.1.0"
pydantic = {extras = ["email"], version = "^1.8.2"}
pytest-asyncio = "^0.15.1"
SQLAlchemy-Utils = "^0.37.8"
starlette = "^0.16.0"
uvicorn = "^0.15.0"
Werkzeug = "^2.0.1"
python-dotenv = "^0.19.0"
websockets = "^10.0"
asyncpg = "^0.24.0"
gunicorn = "^20.1.0"
greenlet = "^1.1.1"
watchdog = "^2.1.5"
idna = "^3.2"
uvloop = "^0.16.0"
watchgod = "^0.7"
password-strength = "^0.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
